test: Move host firewall tests to their own file
